refactor(clients): rename base.py → http.py
File now matches its contents — HTTPClient lives in http.py.

  clients/
      __init__.py   ← re-exports Client + backward-compat aliases
      client.py     ← Client (root)
      http.py       ← HTTPClient (httpx transport)
      sql.py        ← SQLClient (SQLAlchemy transport)
      auth.py       ← AuthStrategy protocol
      auth_strategies/
      azure/        ← AzureClient
